GitWorkingTreeFunTest: Move some test to `GitTest`
Move tests that do not require a working tree to `GitTest`.
